// typechain-types/GBVReportRegistry.ts
export interface GBVReportRegistryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "getReport"
      | "getUserReports"
      | "reports"
      | "submitReport"
      | "userReports"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "ReportSubmitted"): EventFragment;

  encodeFunctionData(
    functionFragment: "getReport",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "getUserReports",
    values: [AddressLike]
  ): string;
  encodeFunctionData(functionFragment: "reports", values: [BytesLike]): string;
  encodeFunctionData(
    functionFragment: "submitReport",
    values: [BytesLike, string[]]
  ): string;
  encodeFunctionData(
    functionFragment: "userReports",
    values: [AddressLike, BigNumberish]
  ): string;

  decodeFunctionResult(functionFragment: "getReport", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "getUserReports",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "reports", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "submitReport",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "userReports",
    data: BytesLike
  ): Result;
}

export namespace ReportSubmittedEvent {
  export type InputTuple = [
    submitter: AddressLike,
    reportHash: BytesLike,
    ipfsCIDs: string[],
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    submitter: string,
    reportHash: string,
    ipfsCIDs: string[],
    timestamp: bigint
  ];
  export interface OutputObject {
    submitter: string;
    reportHash: string;
    ipfsCIDs: string[];
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface GBVReportRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): GBVReportRegistry;
  waitForDeployment(): Promise<this>;

  interface: GBVReportRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getReport: TypedContractMethod<
    [_reportHash: BytesLike],
    [
      [string, string[], bigint, string] & {
        reportHash: string;
        ipfsCIDs: string[];
        timestamp: bigint;
        submitter: string;
      }
    ],
    "view"
  >;

  getUserReports: TypedContractMethod<[_user: AddressLike], [string[]], "view">;

  reports: TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, string] & {
        reportHash: string;
        timestamp: bigint;
        submitter: string;
      }
    ],
    "view"
  >;

  submitReport: TypedContractMethod<
    [_reportHash: BytesLike, _ipfsCIDs: string[]],
    [void],
    "nonpayable"
  >;

  userReports: TypedContractMethod<
    [arg0: AddressLike, arg1: BigNumberish],
    [string],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getReport"
  ): TypedContractMethod<
    [_reportHash: BytesLike],
    [
      [string, string[], bigint, string] & {
        reportHash: string;
        ipfsCIDs: string[];
        timestamp: bigint;
        submitter: string;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "getUserReports"
  ): TypedContractMethod<[_user: AddressLike], [string[]], "view">;
  getFunction(
    nameOrSignature: "reports"
  ): TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, string] & {
        reportHash: string;
        timestamp: bigint;
        submitter: string;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "submitReport"
  ): TypedContractMethod<
    [_reportHash: BytesLike, _ipfsCIDs: string[]],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "userReports"
  ): TypedContractMethod<
    [arg0: AddressLike, arg1: BigNumberish],
    [string],
    "view"
  >;

  getEvent(
    key: "ReportSubmitted"
  ): TypedContractEvent<
    ReportSubmittedEvent.InputTuple,
    ReportSubmittedEvent.OutputTuple,
    ReportSubmittedEvent.OutputObject
  >;

  filters: {
    "ReportSubmitted(address,bytes32,string[],uint256)": TypedContractEvent<
      ReportSubmittedEvent.InputTuple,
      ReportSubmittedEvent.OutputTuple,
      ReportSubmittedEvent.OutputObject
    >;
    ReportSubmitted: TypedContractEvent<
      ReportSubmittedEvent.InputTuple,
      ReportSubmittedEvent.OutputTuple,
      ReportSubmittedEvent.OutputObject
    >;
  };
}